Guide users on the functionality of computers.
The goal of my website will be to give people who are new to computers or do not know much about computers insight into the amount of effort and commitment it takes to build one. It could also act as a guide for those who are interested in building a computer or are thinking of getting into the field.